fix: more pixelated and accurate path
chore: structured code, adding pathfinding service to contentview
