--- IndoorMappingNav/Package/MallMap/Package.realitycomposerpro/PluginData/CF1064AA-49F7-4917-9EB0-5A96E7D1DA98/ShaderGraphEditorPluginID/ShaderGraphEditorPluginID
@@ -1,4 +1,0 @@
-{
-  "materialPreviewObjectType" : 0,
-  "materialPreviewEnvironmentType" : 2
-}
