+++ IndoorMappingNav/Package/MallMap/Package.realitycomposerpro/WorkspaceData/SceneMetadataList.json
@@ -37,8 +37,7 @@
       [
         "801BB739-7548-43D3-AD07-9343D9CB54CB",
         "root",
-        "Timeline",
-        "Track"
+        "Tinesco"
       ],
       {
         "isExpanded" : true,
@@ -48,15 +47,6 @@
         "801BB739-7548-43D3-AD07-9343D9CB54CB",
         "root",
         "Cylinder"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
-      },
-      [
-        "801BB739-7548-43D3-AD07-9343D9CB54CB",
-        "root",
-        "Tinesco"
       ],
       {
         "isExpanded" : true,
@@ -76,6 +66,24 @@
         "root",
         "Timeline",
         "Track2"
+      ],
+      {
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "801BB739-7548-43D3-AD07-9343D9CB54CB",
+        "root"
+      ],
+      {
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "801BB739-7548-43D3-AD07-9343D9CB54CB",
+        "root",
+        "Timeline",
+        "Track"
       ],
       {
         "isExpanded" : true,
@@ -108,14 +116,6 @@
         "disallowsFraming" : false,
         "isExpanded" : true,
         "isLocked" : false
-      },
-      [
-        "801BB739-7548-43D3-AD07-9343D9CB54CB",
-        "root"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
       }
     ]
   },
@@ -123,17 +123,8 @@
     "objectMetadataList" : [
       [
         "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
-        "Root",
-        "SMS_Lt_2_3D_Mapping"
-      ],
-      {
-        "disallowsFraming" : false,
-        "isExpanded" : true,
-        "isLocked" : false
-      },
-      [
-        "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
-        "SMS_1_LT_UG_3D_Mapping_Final"
+        "SMS_1_LT_UG_3D_Mapping_Alpha_Material__1_",
+        "Materials"
       ],
       {
         "isExpanded" : true,
@@ -141,8 +132,26 @@
       },
       [
         "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
-        "SMS_1_LT_UG_3D_Mapping__GLB",
+        "SMS_1_LT_UG_3D_Alpha_Material_jarak_lebih_gede",
         "Geom"
+      ],
+      {
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
+        "Root",
+        "SMS_Lt_2_3D_Mapping",
+        "Plane"
+      ],
+      {
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
+        "SMS_Lt_2_3D_Mapping__3_"
       ],
       {
         "isExpanded" : true,
@@ -159,14 +168,6 @@
       },
       [
         "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
-        "SMS_Lt_2_3D_Mapping__3_"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
-      },
-      [
-        "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
         "SMS_1_LT_UG_3D_Mapping_Final",
         "Geom"
       ],
@@ -176,9 +177,53 @@
       },
       [
         "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
+        "SMS_1_LT_UG_3D_Mapping_Alpha_Material__1_",
+        "Geom"
+      ],
+      {
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
+        "SMS_1_LT_UG_3D_Mapping_Alpha_Material",
+        "Materials"
+      ],
+      {
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
         "Root",
-        "SMS_Lt_2_3D_Mapping",
-        "Plane"
+        "SMS_Lt_2_3D_Mapping"
+      ],
+      {
+        "disallowsFraming" : false,
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
+        "SMS_1_LT_UG_3D_Mapping__GLB",
+        "Geom"
+      ],
+      {
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
+        "SMS_1_LT_UG_3D_Mapping_Alpha_Material",
+        "Geom"
+      ],
+      {
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "CF1064AA-49F7-4917-9EB0-5A96E7D1DA98",
+        "SMS_1_LT_UG_3D_Alpha_Material_jarak_lebih_gede"
       ],
       {
         "isExpanded" : true,
@@ -192,7 +237,7 @@
         "EFA93561-04AB-4694-BE18-5B95DFF15B9F",
         "Root",
         "SMS_Lt_2_3D_Mapping",
-        "Star"
+        "Stroberi"
       ],
       {
         "isExpanded" : true,
@@ -200,9 +245,7 @@
       },
       [
         "EFA93561-04AB-4694-BE18-5B95DFF15B9F",
-        "Root",
-        "SMS_Lt_2_3D_Mapping",
-        "Cube_012"
+        "Root"
       ],
       {
         "isExpanded" : true,
@@ -220,7 +263,8 @@
       [
         "EFA93561-04AB-4694-BE18-5B95DFF15B9F",
         "Root",
-        "SMS_1_LT_UG_3D_Mapping_"
+        "SMS_Lt_2_3D_Mapping",
+        "Star"
       ],
       {
         "isExpanded" : true,
@@ -247,7 +291,18 @@
       },
       [
         "EFA93561-04AB-4694-BE18-5B95DFF15B9F",
-        "Root"
+        "Root",
+        "SMS_1_LT_UG_3D_Mapping_"
+      ],
+      {
+        "isExpanded" : true,
+        "isLocked" : false
+      },
+      [
+        "EFA93561-04AB-4694-BE18-5B95DFF15B9F",
+        "Root",
+        "SMS_Lt_2_3D_Mapping",
+        "Cube_012"
       ],
       {
         "isExpanded" : true,
@@ -258,16 +313,6 @@
         "Root",
         "SMS_1_LT_UG_3D_Mapping_",
         "Geom"
-      ],
-      {
-        "isExpanded" : true,
-        "isLocked" : false
-      },
-      [
-        "EFA93561-04AB-4694-BE18-5B95DFF15B9F",
-        "Root",
-        "SMS_Lt_2_3D_Mapping",
-        "Stroberi"
       ],
       {
         "isExpanded" : true,
